+++ examples/join.php
@@ -6,11 +6,11 @@ use FQL\Query;
 require __DIR__ . '/bootstrap.php';
 
 try {
-    $users = Query\Provider::fromFileQuery('(./examples/data/users.json).data.users')
+    $users = Query\Provider::fromFileQuery('json(./examples/data/users.json).data.users')
         ->select('id, name')
         ->select('o.id')->as('orderId')
         ->select('o.total_price')->as('totalPrice')
-        ->leftJoin(Query\Provider::fromFileQuery('(./examples/data/orders.xml).orders.order'), 'o')
+        ->leftJoin(Query\Provider::fromFileQuery('xml(./examples/data/orders.xml).orders.order'), 'o')
             ->on('id', Operator::EQUAL, 'user_id')
         ->groupBy('o.id')
         ->orderBy('totalPrice')->desc();
@@ -25,9 +25,9 @@ SELECT
     name,
     o.id AS orderId,
     o.total_price AS totalPrice
-FROM [json](./examples/data/users.json).data.users
+FROM json(./examples/data/users.json).data.users
 LEFT JOIN
-    (./examples/data/orders.xml).orders.order AS o
+    xml(./examples/data/orders.xml).orders.order AS o
         ON id = user_id
 GROUP BY o.id
 ORDER BY totalPrice DESC